# pos-products/src/main/java/com/micropos/products/web/ProductsResource.java
package com.micropos.products.web;

import com.micropos.products.api.ProductsApi;
import com.micropos.products.dto.ProductDto;
import com.micropos.products.mapper.ProductMapper;
import com.micropos.products.service.ProductService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

import java.util.ArrayList;
import java.util.List;

@RestController
@RequestMapping("api")
public class ProductsResource implements ProductsApi {

    private final ProductMapper productMapper;

    private final ProductService productService;

    public ProductsResource(ProductService productService, ProductMapper productMapper) {
        this.productMapper = productMapper;
        this.productService = productService;
    }

    @Override
    @RequestMapping(value = "products", method = RequestMethod.GET, produces = "application/json")
    public ResponseEntity<List<ProductDto>> listProducts() {
        List<ProductDto> products = new ArrayList<>(productMapper.toProductsDto(this.productService.products()));
        if (products.isEmpty()) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
        return new ResponseEntity<>(products, HttpStatus.OK);
    }

    @Override
    @RequestMapping(value = "products/{productId}", method = RequestMethod.GET, produces = "application/json")
    public ResponseEntity<ProductDto> showProductById(@PathVariable String productId) {
        ProductDto product = productMapper.toProductDto(productService.getProduct(productId));
        return new ResponseEntity<>(product, HttpStatus.OK);
    }

}